requirements.txt
flask==3.0.0
opencv-python==4.9.0.80
numpy>=1.26.0,<2.0
pillow>=10.2.0
python-dotenv==1.0.0
flask-login==0.6.3
werkzeug==3.0.1
requests>=2.31.0
